% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_plot.R
\name{Sankey_CSMC}
\alias{Sankey_CSMC}
\title{Plot the Sankey plot in ggplot2}
\usage{
Sankey_CSMC(df = df, title = NULL)
}
\arguments{
\item{df}{a dataframe contains the variables Year for x axis, Value for Risk
Contributors, and RiskFactors for different groups.}

\item{title}{optional, to add title to the plot. Default is NULL}
}
\value{
A ggplot object of the Sankey plot
}
\description{
This function generate the Sankey plot with options to save
the plot in user-defined directory.
}
\examples{
data(data_example)
sankey_test = Sankey_CSMC(data_example, title = "Example: Risk Factors for Stroke")

}
